+++ visidata/zscroll.py
@@ -16,3 +16,5 @@ command('ze', 'tmp = cursorVisibleColIndex; pageLeft(); sheet.cursorVisibleColIn
 
 command('zKEY_END',  'sheet.cursorRowIndex = len(rows)-1; sheet.cursorVisibleColIndex = len(visibleCols)-1', 'go to last row and last column')
 command('zKEY_HOME', 'sheet.topRowIndex = sheet.cursorRowIndex = 0; sheet.leftVisibleColIndex = sheet.cursorVisibleColIndex = 0', 'go to top row and top column')
+
+#alias('0', 'gh')